+++ src/components/spartan/SPaginator/SPaginator.vue
@@ -17,13 +17,33 @@ const {
     total,
     canGoPrev = undefined,
     canGoNext = undefined,
+    laravel,
+    inertiaRouter,
 } = defineProps<TPaginatorProps>();
 
 const { t } = translator('dataTable');
 
-const vCount = computed(() => count || (total && size && Math.ceil(total / size)));
+const vQuantity = computed(() => Number(paginatorSize) || 0);
+
+const vTotal = computed(() => {
+    if (total) return total;
+    if (laravel?.total) return laravel.total;
+});
+
+const vSize = computed(() => {
+    if (size) return size;
+    if (laravel?.per_page) return laravel.per_page;
+});
+
+const vCount = computed(() => {
+    if (count) return count;
+    if (laravel?.last_page) return laravel.last_page;
+
+    if (vTotal.value && vSize.value) return Math.ceil(vTotal.value / vSize.value);
+});
+
 const vPageSizes = computed(() => {
-    if (!pageSizes || !size) return;
+    if (!pageSizes || !vSize.value) return;
 
     let data = [...pageSizes];
 
@@ -39,37 +59,41 @@ const vPageSizes = computed(() => {
         }
     }
 
-    if (!data.includes(size)) {
-        data.push(size);
+    if (!data.includes(vSize.value)) {
+        data.push(vSize.value);
         data.sort((a, b) => a - b);
     }
 
     return data;
 });
 
-const quantity = computed(() => Number(paginatorSize) || 0);
+const vPage = computed(() => {
+    if (page) return page;
+
+    if (laravel?.current_page) return laravel.current_page;
+});
 
 const vPages = computed(() => {
-    if (!vCount.value || !page) return [];
+    if (!vCount.value || !vPage.value) return [];
 
     let arr: (string | number)[] = [];
 
-    if (vCount.value <= quantity.value * 2 + 5) return vCount.value;
+    if (vCount.value <= vQuantity.value * 2 + 5) return vCount.value;
 
-    if (page - quantity.value < 4) {
-        arr = Array.apply(null, Array(quantity.value * 2 + 3)).map((_, i) => i + 1);
+    if (vPage.value - vQuantity.value < 4) {
+        arr = Array.apply(null, Array(vQuantity.value * 2 + 3)).map((_, i) => i + 1);
         arr.push('...');
         arr.push(vCount.value);
-    } else if (vCount.value - page - quantity.value < 3) {
+    } else if (vCount.value - vPage.value - vQuantity.value < 3) {
         arr.push(1);
         arr.push('...');
         arr = arr.concat(
-            Array.apply(null, Array(quantity.value * 2 + 3)).map((_, i) => vCount.value! - quantity.value * 2 - 2 + i),
+            Array.apply(null, Array(vQuantity.value * 2 + 3)).map((_, i) => vCount.value! - vQuantity.value * 2 - 2 + i),
         );
     } else {
         arr.push(1);
         arr.push('...');
-        arr = arr.concat(Array.apply(null, Array(1 + quantity.value * 2)).map((_, i) => page - quantity.value + i));
+        arr = arr.concat(Array.apply(null, Array(1 + vQuantity.value * 2)).map((_, i) => vPage.value! - vQuantity.value + i));
         arr.push('...');
         arr.push(vCount.value);
     }
@@ -79,34 +103,62 @@ const vPages = computed(() => {
 
 const vCanGoPrev = computed(() => {
     if (canGoPrev !== undefined) return canGoPrev;
-    return Boolean(page && page > 1);
+    return Boolean(vPage.value && vPage.value > 1);
 });
 
 const vCanGoNext = computed(() => {
     if (canGoNext !== undefined) return canGoNext;
-    return Boolean(page && vCount.value && page < vCount.value);
+    return Boolean(vPage.value && vCount.value && vPage.value < vCount.value);
 });
 
 const prev = () => {
-    if (!page || page === 1) {
+    if (laravel?.prev_page_url) {
+        if (inertiaRouter) {
+            inertiaRouter.visit(laravel.prev_page_url);
+        } else {
+            window.location.href = laravel.prev_page_url;
+        }
+        return;
+    }
+
+    if (!vPage.value || vPage.value === 1) {
         emit('change', { dir: 'prev' });
         return;
     }
-    emit('change', { page: page - 1, dir: 'prev' });
-    emit('update:page', page - 1);
+    emit('change', { page: vPage.value - 1, dir: 'prev' });
+    emit('update:page', vPage.value - 1);
 };
 
 const next = () => {
-    if (!page || page === vCount.value) {
+    if (laravel?.next_page_url) {
+        if (inertiaRouter) {
+            inertiaRouter.visit(laravel.next_page_url);
+        } else {
+            window.location.href = laravel.next_page_url;
+        }
+        return;
+    }
+
+    if (!vPage.value || vPage.value === vCount.value) {
         emit('change', { dir: 'next' });
         return;
     }
-    emit('change', { page: page + 1, dir: 'next' });
-    emit('update:page', page + 1);
+    emit('change', { page: vPage.value + 1, dir: 'next' });
+    emit('update:page', vPage.value + 1);
 };
 
 const selectPage = (pageItem: number) => {
-    if (pageItem === page) return;
+    const link = laravel?.links?.find((link) => link.label === pageItem.toString());
+    if (link?.url) {
+        if (inertiaRouter) {
+            inertiaRouter.visit(link.url);
+        } else {
+            window.location.href = link.url;
+        }
+        return;
+    }
+
+    if (pageItem === vPage.value) return;
     emit('change', { page: pageItem });
     emit('update:page', pageItem);
 };
@@ -122,7 +174,7 @@ const selectPage = (pageItem: number) => {
 
             <SSelect
                 class="h-8 text-xs"
-                :model-value="size"
+                :model-value="vSize"
                 @update:model-value="
                     (value) => {
                         emit('change', { size: Number(value), page: 1 });
@@ -151,7 +203,7 @@ const selectPage = (pageItem: number) => {
                         v-for="pageItem in vPages"
                         :key="pageItem"
                         class="px-4"
-                        :active="Number(pageItem) === page"
+                        :active="Number(pageItem) === vPage"
                         :class="String(pageItem) === '...' ? 'pointer-events-none' : ''"
                         @click="() => Number(pageItem) && selectPage(Number(pageItem))"
                     >