+++ src/components/spartan/SPaginator/types.ts
@@ -11,6 +11,26 @@ export type TPaginatorProps = {
     canGoPrev?: boolean;
     canGoNext?: boolean;
     fitSize?: boolean;
+    inertiaRouter?: any;
+    laravel?: {
+        total?: number;
+        per_page?: number;
+        current_page?: number;
+        last_page?: number;
+        first_page_url?: string | null;
+        last_page_url?: string | null;
+        next_page_url?: string | null;
+        prev_page_url?: string | null;
+        path?: string;
+        from?: number;
+        to?: number;
+        links?: {
+            url: string | null;
+            label: string;
+            active?: boolean;
+        }[];
+        data?: any[];
+    };
 };
 
 export type TPaginatorEmits = {